+++ adapters/yaphooks.pl
@@ -6,7 +6,7 @@
 %  integration code for YAP 6.3.4 and later versions to improve
 %  usability when using the YAP profilers
 %
-%  Last updated on April 25, 2014
+%  Last updated on June 29, 2014
 %
 %  This program is free software: you can redistribute it and/or modify
 %  it under the terms of the GNU General Public License as published by
@@ -40,7 +40,7 @@ user:prolog_predicate_name(user:'$lgt_obj_super_call_'(_, _, _), '^^/2 (from obj
 user:prolog_predicate_name(user:'$lgt_ctg_super_call_'(_, _, _), '^^/2 (from ctg; same pred)') :- !.
 user:prolog_predicate_name(user:'$lgt_call_in_this'(_, _), 'call/1') :- !.
 
-user:prolog_predicate_name(user:'$lgt_send_to_obj_rt'(_, _, _, _, _), '::/2 (runtime)') :- !.
+user:prolog_predicate_name(user:'$lgt_send_to_obj_rt'(_, _, _, _), '::/2 (runtime)') :- !.
 user:prolog_predicate_name(user:'$lgt_send_to_obj'(_, _, _), '::/2 (event-aware)') :- !.
 user:prolog_predicate_name(user:'$lgt_send_to_obj_ne'(_, _, _), '::/2 (event transparent)') :- !.
 user:prolog_predicate_name(user:'$lgt_send_to_self'(_, _, _), '::/1') :- !.